+++ mespronos_group/src/Form/GroupDeleteForm.php
@@ -11,5 +11,33 @@ use Drupal\Core\Entity\ContentEntityDeleteForm;
  */
 class GroupDeleteForm extends ContentEntityDeleteForm {
 
+  /**
+   * {@inheritdoc}
+   */
+  protected function getDeletionMessage() {
+    $entity = $this->getEntity();
+    return $this->t('Le groupe  %label a été supprimé.', [
+      '%label' => $entity->label(),
+    ]);
+  }
+
+  public function getCancelUrl() {
+    /** @var \Drupal\Core\Entity\ContentEntityInterface $entity */
+    $entity = $this->getEntity();
+    return $entity->urlInfo('canonical');
+  }
+
+  /**
+   * {@inheritdoc}
+   */
+  public function getQuestion() {
+    /** @var \Drupal\Core\Entity\ContentEntityInterface $entity */
+    $entity = $this->getEntity();
+
+    return $this->t('Êtes-vous sûr de vouloir supprimer le groupe %label?', [
+      '%label' => $this->getEntity()->label(),
+    ]);
+
+  }
 
 }